# Optimise `ImageColor` using `functools.lru_cache`